cuddle/requirements.txt
asgiref==3.5.2
Django==4.0.6
django-countries==7.4.2
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.4.0
tzdata==2022.1
